moved class Node from InMemoryHistoryManager
