+++ src/manager/Node.java
@@ -1,0 +1,33 @@
+package manager;
+
+import task.Task;
+
+class Node {
+	private Task task;
+	private Node prev;
+	private Node next;
+
+	public Node getNext() {
+		return next;
+	}
+
+	public Node getPrev() {
+		return prev;
+	}
+
+	public Task getTask() {
+		return task;
+	}
+
+	public void setNext(Node next) {
+		this.next = next;
+	}
+
+	public void setPrev(Node prev) {
+		this.prev = prev;
+	}
+
+	public void setTask(Task task) {
+		this.task = task;
+	}
+}
